feat: implement change analysis utility for editor state comparison
